refactor(landing): split page into focused section components
